fastapi_backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
supabase==2.3.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
asyncpg==0.29.0
bluetooth-py==0.1.0
pyserial==3.5
